+++ anyblok/model/hybrid_method.py
@@ -60,6 +60,9 @@ class HybridMethodPlugin(ModelPluginBase):
                 if self is self_:
                     return getattr(super(new_base, self), attr)(
                         self, *args, **kwargs)
+                elif hasattr(self, '_aliased_insp'):
+                    return getattr(super(new_base, self._aliased_insp._target),
+                                   attr)(self, *args, **kwargs)
                 else:
                     return getattr(super(new_base, self), attr)(
                         *args, **kwargs)