+++ anyblok/tests/testcase.py
@@ -1,9 +1,12 @@
 # This file is a part of the AnyBlok project
 #
 #    Copyright (C) 2018 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
+#    Copyright (C) 2019 Jean-Sebastien SUZANNE <jssuzanne@anybox.fr>
 #
 # This Source Code Form is subject to the terms of the Mozilla Public License,
 # v. 2.0. If a copy of the MPL was not distributed with this file,You can
 # obtain one at http://mozilla.org/MPL/2.0/.
+
 from ..testing import (  # noqa
-    TestCase, DBTestCase, BlokTestCase, SharedDataTestCase, LogCapture)
+    TestCase, DBTestCase, BlokTestCase, SharedDataTestCase, LogCapture,
+    skip_unless_bloks_installed)